+++ src/components/menu/toolbar/Settings.tsx
@@ -3,16 +3,19 @@ import {
   Cog6ToothIcon,
   RectangleStackIcon,
   EllipsisHorizontalIcon,
+  ArrowPathIcon,
+  LinkIcon,
 } from '@heroicons/react/24/outline';
 import debugModule from 'debug';
 import { clearCache } from '../../../DataManager';
 import { AlertData } from '../../../types/AlertTypes';
-import { PlanSpecialFunctions } from '../../../types/PlanTypes';
 import Utility from '../../../utility/Utility';
+import { LoadLegacyUrlFunction } from '../../../types/BaseTypes';
+import { toast } from 'react-hot-toast';
 
-const settingsMenu = (f2: PlanSpecialFunctions): AlertData => ({
+const settingsMenu = (loadLegacyUrl: LoadLegacyUrlFunction): AlertData => ({
   title: 'Settings',
-  message: `Customize your Paper experience with! These settings are saved in your browser and not in the URL.`,
+  message: `Customize your Paper experience! These settings are saved in your browser and not in the URL.`,
   cancelButton: 'Close',
   color: 'orange',
   icon: Cog6ToothIcon,
@@ -45,22 +48,61 @@ const settingsMenu = (f2: PlanSpecialFunctions): AlertData => ({
             },
           },
           {
-            switch: 'use_url',
-            title: 'Save data to URL',
+            title: 'Clear local course data cache',
             description:
-              "By default, your plan and schedule data is saved to the URL. However, that can get a bit annoying if your browser keeps routing to an old link. If you just use your account or browser storage to save your plans and schedules, you can disable URL saving. You won't be able to open Paper links while this setting is disabled unless they're shortened.",
-            saveToStorage: true,
+              'This clears all of the course data stored in your browser without deleting any plans or schedules.',
+            action: (_, next) => {
+              next({
+                title: 'Clear local course data cache?',
+                message:
+                  'All of the course data stored in your browser will be erased. Any plans or schedules will not be affected, including ones saved in your browser. The site will immediately reload to get the latest data.',
+                color: 'red',
+                icon: ArrowPathIcon,
+                cancelButton: 'Cancel',
+                confirmButton: 'Clear',
+                action: () => {
+                  clearCache().finally(() => {
+                    window.location.reload();
+                  });
+                },
+              });
+            },
           },
-          // {
-          //   title: 'Clear local course data cache',
-          //   description:
-          //     'This clears all of the course data stored in your browser. Clearing the data cache will not delete any plans or schedules, including ones saved in your browser. THE SITE WILL RELOAD!',
-          //   singleAction: () => {
-          //     clearCache().finally(() => {
-          //       window.location.reload();
-          //     });
-          //   },
-          // },
+          {
+            title: 'Load legacy URL',
+            description:
+              'Load old Paper plan and schedule URLs (or Plan Northwestern plan URLs).',
+            action: (_, next) => {
+              next({
+                title: 'Load legacy URL',
+                color: 'purple',
+                icon: LinkIcon,
+                message: 'Enter the legacy plan or schedule URL below.',
+                textInput: {
+                  focusByDefault: true,
+                  placeholder: 'https://www.paper.nu/?...',
+                  match:
+                    /https?:\/\/(?:www\.)?(?:paper\.nu|plan-nu\.com)\/?\?[^#]+$/,
+                  matchError: 'Invalid legacy plan or schedule URL',
+                },
+                cancelButton: 'Cancel',
+                confirmButton: 'Load',
+                action: ({ inputText }) => {
+                  if (!inputText) {
+                    return;
+                  }
+
+                  try {
+                    const url = new URL(inputText);
+                    loadLegacyUrl(url);
+                  } catch (e) {
+                    console.error(e);
+                    toast.error('Invalid legacy plan or schedule URL');
+                  }
+                },
+              });
+            },
+          },
         ],
       },
       {