+++ src/components/plan/Content.tsx
@@ -1,4 +1,4 @@
-import { PlusIcon } from '@heroicons/react/24/outline';
+import { PlusIcon, TrashIcon } from '@heroicons/react/24/outline';
 import React from 'react';
 import PlanManager from '../../PlanManager';
 import { Alert } from '../../types/AlertTypes';
@@ -12,6 +12,7 @@ import Utility from '../../utility/Utility';
 import Year from './Year';
 import { motion } from 'framer-motion';
 import { SideCard } from '../../types/SideCardTypes';
+import PlanTaskbarButton from './PlanTaskbarButton';
 
 interface ContentProps {
   data: PlanData;
@@ -59,34 +60,52 @@ class Content extends React.Component<ContentProps> {
         transition={{ duration: 0.25 }}
       >
         {years}
-        <div className="m-5 flex justify-center gap-4">
+        <div className="m-5 flex flex-col items-center justify-center gap-4 sm:flex-row">
           <div className="w-48 rounded-lg border-2 border-gray-200 p-1 shadow-sm dark:border-gray-600">
             <p className="text-center text-sm font-light text-gray-400 dark:text-gray-400">
               <span className="font-medium">{units}</span> total {unitString}
             </p>
           </div>
-          {content.courses.length < 10 && (
-            <button
-              className="block rounded-lg bg-gray-200 px-5 py-1 text-gray-400 shadow-sm
-                            hover:bg-gray-300 hover:text-gray-500 dark:bg-gray-700 dark:hover:bg-gray-600 dark:hover:text-gray-300"
+          <div className="flex gap-4">
+            {content.courses.length < 10 && (
+              <PlanTaskbarButton
+                onClick={() => {
+                  this.props.alert({
+                    title: 'Add a year?',
+                    message:
+                      'This will add another year to your plan. You can remove it by removing all classes from that year and refreshing the page.',
+                    confirmButton: 'Add year',
+                    cancelButton: 'Close',
+                    color: 'cyan',
+                    icon: PlusIcon,
+                    action: () => {
+                      this.props.f2.addYear();
+                    },
+                  });
+                }}
+              >
+                Add year
+              </PlanTaskbarButton>
+            )}
+            <PlanTaskbarButton
               onClick={() => {
                 this.props.alert({
-                  title: 'Add a year?',
+                  title: 'Clear plan?',
                   message:
-                    'This will add another year to your plan. You can remove it by removing all classes from that year and refreshing the page.',
-                  confirmButton: 'Add year',
-                  cancelButton: 'Close',
-                  color: 'cyan',
-                  icon: PlusIcon,
+                    'All of your current plan data, which includes everything for each year and everything in your bookmarks, will be cleared. Make sure to save any data you want to keep.',
+                  color: 'red',
+                  icon: TrashIcon,
+                  cancelButton: 'Cancel',
+                  confirmButton: 'Clear',
                   action: () => {
-                    this.props.f2.addYear();
+                    this.props.f2.clearData();
                   },
                 });
               }}
             >
-              Add year
-            </button>
-          )}
+              Clear plan
+            </PlanTaskbarButton>
+          </div>
         </div>
       </motion.div>
     );